Place unit tests about getters on the base transaction entity file
