+++ framework/test/mocha/unit/components/storage/entities/transaction.js
@@ -25,10 +25,34 @@ const transactionsFixtures = require('../../../../fixtures').transactions;
 const {
 	NonSupportedFilterTypeError,
 	NonSupportedOperationError,
+	NonSupportedOptionError,
 } = require('../../../../../../src/components/storage/errors');
 
 const numSeedRecords = 5;
 const NON_EXISTENT_ID = '1234';
+
+const expectValidTransaction = (result, transaction, extended = true) => {
+	expect(result.id).to.be.eql(transaction.id);
+	expect(result.senderId).to.be.eql(transaction.senderId);
+	expect(result.recipientId).to.be.eql(transaction.recipientId);
+	expect(result.senderPublicKey).to.be.eql(transaction.senderPublicKey);
+
+	// As we directly saved the transaction and not applied the account
+	// So the recipientPublicKey for the account is not updated
+	// expect(result.recipientPublicKey).to.be.eql(transaction.recipientPublicKey);
+
+	expect(result.requesterPublicKey).to.be.eql(transaction.requesterPublicKey);
+	expect(result.signature).to.be.eql(transaction.signature);
+	expect(result.signatures).to.be.eql(transaction.signatures);
+	expect(result.amount).to.be.eql(transaction.amount);
+	expect(result.fee).to.be.eql(transaction.fee);
+	expect(result.timestamp).to.be.eql(transaction.timestamp);
+	expect(result.type).to.be.eql(transaction.type);
+
+	if (extended) {
+		expect(result.asset).to.be.eql(transaction.asset);
+	}
+};
 
 describe('Transaction', () => {
 	let adapter;
@@ -37,6 +61,9 @@ describe('Transaction', () => {
 	let addFieldSpy;
 	let validFilters;
 	let SQLs;
+	let validOptions;
+	let validSimpleObjectFields;
+	let validExtendedObjectFields;
 
 	before(async () => {
 		storage = new storageSandbox.StorageSandbox(
@@ -130,6 +157,50 @@ describe('Transaction', () => {
 			'dapp_link',
 		];
 
+		validSimpleObjectFields = [
+			'id',
+			'blockId',
+			'height',
+			'type',
+			'timestamp',
+			'senderId',
+			'recipientId',
+			'amount',
+			'fee',
+			'signature',
+			'signSignature',
+			'signatures',
+			'senderPublicKey',
+			'recipientPublicKey',
+			'requesterPublicKey',
+			'confirmations',
+		];
+
+		validExtendedObjectFields = [
+			'asset',
+			'id',
+			'blockId',
+			'height',
+			'type',
+			'timestamp',
+			'senderId',
+			'recipientId',
+			'amount',
+			'fee',
+			'signature',
+			'signSignature',
+			'signatures',
+			'senderPublicKey',
+			'recipientPublicKey',
+			'requesterPublicKey',
+			'confirmations',
+		];
+
+		validOptions = {
+			limit: 100,
+			offset: 0,
+		};
+
 		adapter = storage.adapter;
 	});
 
@@ -188,6 +259,361 @@ describe('Transaction', () => {
 		it('should setup specific filters', async () => {
 			const transaction = new Transaction(adapter);
 			expect(transaction.getFilters()).to.have.members(validFilters);
+		});
+	});
+
+	describe('get()', () => {
+		beforeEach(() => {
+			return sinonSandbox.restore();
+		});
+		it('should accept only valid filters', async () => {
+			// Arrange
+			const transaction = new Transaction(adapter);
+			const transactions = [
+				new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				}),
+			];
+			await storage.entities.Transaction.create(transactions);
+			const validFilter = {
+				id: transactions[0].id,
+			};
+			// Act & Assert
+			expect(() => {
+				transaction.get(validFilter);
+			}).not.to.throw(NonSupportedFilterTypeError);
+		});
+		it('should throw error for invalid filters');
+		it('should accept only valid options');
+		it('should throw error for invalid options');
+
+		it('should call adapter.executeFile with proper param for extended=false', async () => {
+			// Arrange
+			sinonSandbox.spy(adapter, 'executeFile');
+			const transaction = new Transaction(adapter);
+			// Act
+			transaction.get();
+			// Assert
+			expect(adapter.executeFile.firstCall.args[0]).to.be.eql(SQLs.select);
+		});
+
+		it('should call adapter.executeFile with proper param for extended=true', async () => {
+			// Arrange
+			sinonSandbox.spy(adapter, 'executeFile');
+			const transaction = new Transaction(adapter);
+			// Act
+			transaction.get({}, { extended: true });
+			// Assert
+			expect(adapter.executeFile.firstCall.args[0]).to.be.eql(
+				SQLs.selectExtended
+			);
+		});
+
+		it('should accept "tx" as last parameter and pass to adapter.executeFile', async () => {
+			// Arrange
+			const transaction = new Transaction(adapter);
+			const transactions = [
+				new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				}),
+			];
+			await storage.entities.Transaction.create(transactions);
+			const validFilter = {
+				id: transactions[0].id,
+			};
+			const _getSpy = sinonSandbox.spy(transaction, 'get');
+			// Act & Assert
+			await transaction.begin('testTX', async tx => {
+				await transaction.get(validFilter, {}, tx);
+				expect(Object.getPrototypeOf(_getSpy.firstCall.args[2])).to.be.eql(
+					Object.getPrototypeOf(tx)
+				);
+			});
+		});
+
+		it('should resolve with one object matching specification of type definition of simple object', async () => {
+			// Arrange
+			const transaction = new Transaction(adapter);
+			const transactions = [
+				new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				}),
+			];
+			await storage.entities.Transaction.create(transactions);
+			const validFilter = {
+				id: transactions[0].id,
+			};
+			// Act
+			const results = await transaction.get(validFilter, { extended: false });
+			// Assert
+			expect(results[0]).to.have.all.keys(validSimpleObjectFields);
+		});
+
+		it('should resolve with one object matching specification of type definition of full object', async () => {
+			// Arrange
+			const transaction = new Transaction(adapter);
+			const transactions = [
+				new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				}),
+			];
+			await storage.entities.Transaction.create(transactions);
+			const validFilter = {
+				id: transactions[0].id,
+			};
+			// Act
+			const results = await transaction.get(validFilter, { extended: true });
+			// Assert
+			expect(results[0]).to.have.all.keys(validExtendedObjectFields);
+		});
+		it('should not change any of the provided parameter');
+
+		it('should return result in valid format', async () => {
+			let transaction = null;
+			const transactions = [];
+
+			for (let i = 0; i < numSeedRecords; i++) {
+				transaction = new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				});
+				transactions.push(transaction);
+			}
+			await storage.entities.Transaction.create(transactions);
+
+			// Check for last transaction
+			const result = await storage.entities.Transaction.get({
+				id: transaction.id,
+			});
+
+			expect(result).to.not.empty;
+			expect(result).to.be.lengthOf(1);
+			expect(result[0].id).to.be.eql(transaction.id);
+			expectValidTransaction(result[0], transaction, false);
+		});
+
+		it('should return result valid format for extended version', async () => {
+			let transaction = null;
+			const transactions = [];
+
+			for (let i = 0; i < numSeedRecords; i++) {
+				transaction = new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				});
+				transactions.push(transaction);
+			}
+			await storage.entities.Transaction.create(transactions);
+
+			// Check for last transaction
+			const result = await storage.entities.Transaction.get(
+				{
+					id: transaction.id,
+				},
+				{
+					limit: 10,
+					offset: 0,
+					extended: true,
+				}
+			);
+
+			expect(result).to.not.empty;
+			expect(result).to.be.lengthOf(1);
+			expect(result[0].id).to.be.eql(transaction.id);
+			expectValidTransaction(result[0], transaction);
+		});
+
+		it('should paginate the results properly', async () => {
+			const block = seeder.getLastBlock();
+			let transaction = null;
+			const transactions = [];
+
+			for (let i = 0; i < numSeedRecords; i++) {
+				transaction = new transactionsFixtures.Transaction({
+					blockId: block.id,
+				});
+				transactions.push(transaction);
+			}
+			await storage.entities.Transaction.create(transactions);
+
+			const result1 = await storage.entities.Transaction.get(
+				{
+					blockId: block.id,
+				},
+				{
+					limit: 2,
+					offset: 0,
+				}
+			);
+
+			const result2 = await storage.entities.Transaction.get(
+				{
+					blockId: block.id,
+				},
+				{
+					limit: 2,
+					offset: 1,
+				}
+			);
+			expect(result1).to.not.empty;
+			expect(result2).to.not.empty;
+			expect(result1).to.be.lengthOf(2);
+			expect(result2).to.be.lengthOf(2);
+			return expect(result1[1]).to.be.eql(result2[0]);
+		});
+
+		it('should sort the results for provided "sortField"', async () => {
+			const block = seeder.getLastBlock();
+			const transactions = [];
+
+			for (let i = 0; i < numSeedRecords; i++) {
+				transactions.push(
+					new transactionsFixtures.Transaction({
+						blockId: block.id,
+					})
+				);
+			}
+			await storage.entities.Transaction.create(transactions);
+
+			const result = await storage.entities.Transaction.get(
+				{
+					blockId: block.id,
+				},
+				{
+					sort: 'id:desc',
+					limit: 20,
+					offset: 0,
+				}
+			);
+
+			expect(result).to.be.eql(_.orderBy(result, 'id', 'desc'));
+		});
+
+		describe('filters', () => {
+			// To make add/remove filters we add their tests.
+			it('should have only specific filters', async () => {
+				const transaction = new Transaction(adapter);
+				expect(transaction.getFilters()).to.eql(validFilters);
+			});
+			// For each filter type
+			it('should return matching result for provided filter');
+		});
+	});
+
+	describe('getOne()', () => {
+		it('should accept only valid filters', async () => {
+			// Arrange
+			const aTransaction = new transactionsFixtures.Transaction({
+				blockId: seeder.getLastBlock().id,
+			});
+			const transaction = new Transaction(adapter);
+			await storage.entities.Transaction.create(aTransaction);
+			const validFilter = {
+				senderId: aTransaction.senderId,
+			};
+
+			// Act & Assert
+			expect(() => {
+				transaction.getOne(validFilter);
+			}).not.to.throw(NonSupportedFilterTypeError);
+		});
+
+		// The implementation of the test is ready but should work implementation in code in different PR
+		// eslint-disable-next-line mocha/no-skipped-tests
+		it.skip('should throw error for invalid filters', async () => {
+			// Arrange
+			const aTransaction = new transactionsFixtures.Transaction({
+				blockId: seeder.getLastBlock().id,
+			});
+			const transaction = new Transaction(adapter);
+			await storage.entities.Transaction.create(aTransaction);
+			const invalidFilter = {
+				nonExistentField: aTransaction.senderId,
+			};
+
+			// Act & Assert
+			expect(() => {
+				transaction.getOne(invalidFilter);
+			}).to.throw(NonSupportedFilterTypeError);
+		});
+
+		// The implementation of the test is ready but should work implementation in code in different PR
+		// eslint-disable-next-line mocha/no-skipped-tests
+		it.skip('should accept only valid options', async () => {
+			// Arrange
+			const transaction = new Transaction(adapter);
+			// Act & Assert
+			expect(() => {
+				transaction.getOne({}, validOptions);
+			}).not.to.throw(NonSupportedOptionError);
+		});
+
+		// The implementation of the test is ready but should work implementation in code in different PR
+		// eslint-disable-next-line mocha/no-skipped-tests
+		it.skip('should throw error for invalid options', async () => {
+			// Arrange
+			const aTransaction = new transactionsFixtures.Transaction({
+				blockId: seeder.getLastBlock().id,
+			});
+			const transaction = new Transaction(adapter);
+			await storage.entities.Transaction.create(aTransaction);
+			const invalidOptions = {
+				foo: 'bar',
+			};
+			// Act & Assert
+			expect(() => {
+				transaction.getOne({}, invalidOptions);
+			}).to.throw(NonSupportedOptionError);
+		});
+
+		it('should resolve with one object matching specification of type definition of simple object', async () => {
+			// Arrange
+			const aTransaction = new transactionsFixtures.Transaction({
+				blockId: seeder.getLastBlock().id,
+			});
+			const transaction = new Transaction(adapter);
+			await storage.entities.Transaction.create(aTransaction);
+			// Act
+			const results = await transaction.getOne(
+				{
+					id: aTransaction.id,
+				},
+				{
+					extended: false,
+				}
+			);
+			expect(results).to.have.all.keys(validSimpleObjectFields);
+		});
+
+		it('should reject with error if matched with multiple records for provided filters', async () => {
+			// Arrange
+			const transaction = new Transaction(adapter);
+			const transactions = [
+				new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				}),
+				new transactionsFixtures.Transaction({
+					blockId: seeder.getLastBlock().id,
+				}),
+			];
+			await storage.entities.Transaction.create(transactions);
+			// Act
+			expect(
+				transaction.getOne({
+					blockId: transactions[0].blockId,
+				})
+			).to.be.rejected;
+			// Assert
+		});
+
+		describe('filters', () => {
+			// To make add/remove filters we add their tests.
+			it('should have only specific filters', async () => {
+				// Arrange
+				const transaction = new Transaction(adapter);
+				// Act & Assert
+				expect(transaction.getFilters()).to.eql(validFilters);
+			});
+			// For each filter type
+			it('should return matching result for provided filter');
 		});
 	});
